Version 1.1.2 - fix some minor stuff, remove VOL from stream in mp4creator

+++ server/mp4live/video_v4l2_source.h
@@ -65,7 +65,7 @@ class CV4L2VideoSource : public CMediaSource {
 
   typedef struct {
     void* start;
-    size_t length;
+    __u32 length;
   } capture_buffer_t;
 
   capture_buffer_t* m_buffers;
